@@ -1,21 +1,28 @@
 {
-	"name": "tijsverkoyen/bpost",
-	"type": "library",
-	"description": "PHP bPost is a (wrapper)class to communicate with the bPost Webservices.",
-	"homepage": "https://github.com/tijsverkoyen/bpost",
-	"license": "BSD",
-	"authors": [
-		{
-			"name": "Tijs Verkoyen",
-			"email": "bpost@verkoyen.eu",
-			"role": "Developer"
-		}
-	],
-	"require": {
-		"php": ">=5.3.0",
-		"ext-curl": "*"
-	},
-	"autoload": {
-		"classmap": [""]
-	}
+  "name": "tavy315/bpost",
+  "type": "library",
+  "description": "PHP bPost is a (wrapper)class to communicate with the bPost Webservices.",
+  "homepage": "https://github.com/tavy315/bpost",
+  "license": "BSD",
+  "authors": [
+    {
+      "name": "Tijs Verkoyen",
+      "email": "bpost@verkoyen.eu",
+      "role": "Developer"
+    },
+    {
+      "name": "Octavian Matei",
+      "homepage": "http://www.octav.name",
+      "role": "Contributor"
+    }
+  ],
+  "require": {
+    "php": ">=5.4.0",
+    "ext-curl": "*"
+  },
+  "autoload": {
+    "classmap": [
+      ""
+    ]
+  }
 }
