@@ -26,14 +26,14 @@ class AddressTest extends \PHPUnit_Framework_TestCase
      */
     public function testToXML()
     {
-        $data = array(
-            'streetName' => 'Afrikalaan',
-            'number' => '2890',
-            'box' => '3',
-            'postalCode' => '9000',
-            'locality' => 'Gent',
+        $data = [
+            'streetName'  => 'Afrikalaan',
+            'number'      => '2890',
+            'box'         => '3',
+            'postalCode'  => '9000',
+            'locality'    => 'Gent',
             'countryCode' => 'BE',
-        );
+        ];
 
         $expectedDocument = self::createDomDocument();
         $address = $expectedDocument->createElement('common:address');
@@ -115,14 +115,14 @@ class AddressTest extends \PHPUnit_Framework_TestCase
      */
     public function testCreateFromXML()
     {
-        $data = array(
-            'streetName' => 'Afrikalaan',
-            'number' => '289',
-            'box' => '3',
-            'postalCode' => '9000',
-            'locality' => 'Gent',
+        $data = [
+            'streetName'  => 'Afrikalaan',
+            'number'      => '289',
+            'box'         => '3',
+            'postalCode'  => '9000',
+            'locality'    => 'Gent',
             'countryCode' => 'BE',
-        );
+        ];
 
         $document = self::createDomDocument();
         $addressElement = $document->createElement('address');
