@@ -27,20 +27,20 @@ class SenderTest extends \PHPUnit_Framework_TestCase
      */
     public function testToXML()
     {
-        $data = array(
-            'name' => 'Tijs Verkoyen',
+        $data = [
+            'name'    => 'Tijs Verkoyen',
             'company' => 'Sumo Coders',
-            'address' => array(
-                'streetName' => 'Afrikalaan',
-                'number' => '289',
-                'box' => '3',
-                'postalCode' => '9000',
-                'locality' => 'Gent',
+            'address' => [
+                'streetName'  => 'Afrikalaan',
+                'number'      => '289',
+                'box'         => '3',
+                'postalCode'  => '9000',
+                'locality'    => 'Gent',
                 'countryCode' => 'BE',
-            ),
+            ],
             'emailAddress' => 'bpost@verkoyen.eu',
-            'phoneNumber' => '+32 9 395 02 51',
-        );
+            'phoneNumber'  => '+32 9 395 02 51',
+        ];
 
         $expectedDocument = self::createDomDocument();
         $sender = $expectedDocument->createElement('sender');
