@@ -13,19 +13,19 @@ class PoiTest extends \PHPUnit_Framework_TestCase
      */
     public function testCreateFromXml()
     {
-        $data = array(
-            'Id' => 405700,
-            'Type' => 1,
-            'Name' => 'ST-AMANDSBERG HOGEWEG',
-            'Street' => 'Hogeweg',
-            'Number' => '108',
-            'Zip' => '9040',
-            'City' => 'Sint-Amandsberg',
-            'X' => 106482,
-            'Y' => 195899,
+        $data = [
+            'Id'        => 405700,
+            'Type'      => 1,
+            'Name'      => 'ST-AMANDSBERG HOGEWEG',
+            'Street'    => 'Hogeweg',
+            'Number'    => '108',
+            'Zip'       => '9040',
+            'City'      => 'Sint-Amandsberg',
+            'X'         => 106482,
+            'Y'         => 195899,
             'Longitude' => 3.7479,
-            'Latitude' => 51.0716,
-        );
+            'Latitude'  => 51.0716,
+        ];
 
         // build xml
         $xmlString = '<record>';
