@@ -4,7 +4,6 @@ namespace TijsVerkoyen\Bpost\Bpost\Order\Box;
 
 interface IBox
 {
-
     /**
      * @param array $options
      */
@@ -38,14 +37,17 @@ interface IBox
 
     /**
      * Return the object as an array for usage in the XML
-     * @param  \DomDocument $document
-     * @param  string $prefix
+     *
+     * @param \DomDocument $document
+     * @param string       $prefix
+     *
      * @return \DomElement
      */
     public function toXML(\DOMDocument $document, $prefix = null);
 
     /**
-     * @param  \SimpleXMLElement $xml
+     * @param \SimpleXMLElement $xml
+     *
      * @return self
      */
     public static function createFromXML(\SimpleXMLElement $xml);
