@@ -368,12 +368,13 @@ class Poi
     /**
      * Create a POI based on an XML-object
      *
-     * @param  \SimpleXMLElement $xml
+     * @param \SimpleXMLElement $xml
+     *
      * @return Poi
      */
     public static function createFromXML(\SimpleXMLElement $xml)
     {
-        $poi = new Poi();
+        $poi = new self();
 
         if (isset($xml->Id) && $xml->Id != '') {
             $poi->setId((string) $xml->Id);
